refactor MainNavigation
