--- src/components/MainNavigation.vue
@@ -1,8 +1,0 @@
-<template>
-    <q-bar>
-        <q-btn color="primary" flat :label="$t('home')" @click="$router.push({ name: 'home' })"/>
-    </q-bar>
-</template>
-
-<script setup lang="ts">
-</script>
