+++ src/components/lib/navigation/Navigation.vue
@@ -1,0 +1,8 @@
+<template>
+    <q-bar>
+        <slot/>
+    </q-bar>
+</template>
+
+<script setup lang="ts">
+</script>
