+++ src/pages/CreateSessionPage.vue
@@ -3,5 +3,5 @@
 </template>
 
 <script setup lang="ts">
-import CreateSessionForm from '@/components/session/CreateSessionForm.vue'
+import CreateSessionForm from '@/components/app/session/CreateSessionForm.vue'
 </script>
